+++ tracker/forms.py
@@ -1,4 +1,7 @@
 from django import forms
+from timezone_field import forms as tzforms
+
+from tracker.models import Setting
 
 
 class AddTimeRecordForm(forms.Form):
@@ -12,3 +15,13 @@ class AddTimeRecordForm(forms.Form):
         required=False,
         widget=forms.DateTimeInput(attrs={'type': 'datetime-local', 'class': 'form-control'}))
 
+
+class SettingsForm(forms.ModelForm):
+    timezone = tzforms.TimeZoneFormField(widget=forms.Select(attrs={'class': 'form-control'}))
+    locale = forms.ChoiceField(choices=Setting.LOCALES, widget=forms.Select(attrs={'class': 'form-control'}))
+    duration_format = forms.ChoiceField(choices=Setting.DURATION_FORMATS,
+                                        widget=forms.Select(attrs={'class': 'form-control'}))
+
+    class Meta:
+        model = Setting
+        fields = ('timezone', 'locale', 'duration_format')