+++ tracker/models/project.py
@@ -24,7 +24,10 @@ class Project(models.Model):
             .order_by('-modification_time') \
             .first()
 
-        return format_timespan(timezone.now() - last_time_record.modification_time)
+        if not last_time_record:
+            return 'Never updated'
+
+        return 'Last updated {} ago'.format(format_timespan(timezone.now() - last_time_record.modification_time))
 
     def is_tracking(self, user):
         no_end_time = self.timerecord_set \
@@ -42,6 +45,22 @@ class Project(models.Model):
 
     def is_member(self, user):
         return user in self.admins.all() or user in self.editors.all()
+
+    @staticmethod
+    def recent_projects(user, num=3):
+        user_projects = Project.objects \
+            .filter(timerecord__user=user) \
+            .order_by('-timerecord__modification_time')
+
+        recent_projects = []
+        for i in range(num):
+            p = user_projects \
+                .exclude(id__in=(p.id for p in recent_projects)) \
+                .first()
+            if not p:
+                break
+            recent_projects.append(p)
+        return recent_projects
 
 
 location = os.path.dirname(__file__)