+++ tracker/tests/__init__.py
@@ -1,0 +1,2 @@
+from tracker.tests.models import *
+from tracker.tests.views import *