+++ tracker/tests/models/__init__.py
@@ -1,0 +1,3 @@
+from tracker.tests.models.organization import OrganizationModelTests
+from tracker.tests.models.project import ProjectModelTests
+from tracker.tests.models.time_record import TimeRecordModelTests