+++ tracker/tests/views/__init__.py
@@ -1,0 +1,1 @@
+from tracker.tests.views.user import UserTests